// NutriTEC/nutritec.client/src/app/interfaces/RegisterClient.ts
export interface RegisterClient {
  nombre: string;
  apellido1: string;
  apellido2: string;
  nacimiento: Date;
  peso: number;
  IMC: number;
  pais: string;
  cintura: number;
  cuello: number;
  caderas: number;
  musculo: number;
  grasa: number;
  calorias: number;
  email: string;
  password: string;
}